epg-updaterワーカーの実装

+++ rust/libs/domain/src/events/kurec_events.rs
@@ -1,13 +1,17 @@
 use serde::{Deserialize, Serialize};
+// use shared_core::event_metadata::Event; // #[event] マクロが自動で実装するため不要
+use shared_macros::event; // #[event] マクロをインポート
 
 /// EPG情報がKVSに保存されたことを示すイベント。
 /// 後続のワーカー (例: Meilisearch登録ワーカー) をトリガーするために使用される。
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+// Event トレイトを実装
+#[event(stream = "kurec-epg-updated", subject = "epg.stored")] // 仮のストリーム名とサブジェクトプレフィックス
 pub struct EpgStoredEvent {
     /// 番組情報を取得したmirakcのベースURL
     pub mirakc_url: String,
     /// 更新されたEPGのサービスID (Mirakurun Service ID)
-    pub service_id: i32,
+    pub service_id: i64, // i32 -> i64
 }
 
 #[cfg(test)]
